feat gate and update book

+++ app/Http/Controllers/BookController.php
@@ -2,19 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use File;
-use Exception;
 use App\Models\Book;
-use App\Models\User;
 use App\Models\BookCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use function Laravel\Prompts\error;
 use Illuminate\Support\Facades\Auth;
-use App\Http\Requests\StoreBookRequest;
-
-use Illuminate\Support\Facades\Storage;
-use App\Http\Requests\UpdateBookRequest;
+use Illuminate\Support\Facades\Gate;
 
 class BookController extends Controller
 {
@@ -25,7 +18,9 @@ class BookController extends Controller
      */
     public function index()
     {
-        $books = Book::orderBy('updated_at', 'DESC')->with(['user', 'category'])->get();
+
+        $books = Book::orderBy('updated_at', 'DESC')->where('deleted_at', null)
+            ->with(['user', 'category'])->get();
 
         // role id for admin = 2
         if (Auth::user()->role_id !== 2) {
@@ -55,7 +50,7 @@ class BookController extends Controller
     public function store(Request $request)
     {
         $validate = $request->validate([
-            'title' => 'required|string|max:255',
+            'title' => 'required|string|max:255|unique:books,title',
             'quantity' => 'required|integer',
             'cover' => 'required|mimes:jpg,jpeg,png,bmp,tiff|max:4096',
             'file' => 'required|mimes:pdf|max:10000',
@@ -101,6 +96,10 @@ class BookController extends Controller
     {
         $book = Book::where('slug', $slug)->with(['category', 'user'])->firstOrFail();
 
+        if (!Gate::allows('get-book', $book)) {
+            return to_route('home')->with('warning', 'Anda Tidak Punya Ijin Melihat Data Ini');
+        }
+
         return view('admin.pages.show-book', compact('book'));
     }
 
@@ -113,6 +112,9 @@ class BookController extends Controller
     public function edit(string $slug)
     {
         $book = Book::where('slug', $slug)->with(['user', 'category'])->firstOrFail();
+        if (!Gate::allows('update-book', $book)) {
+            return to_route('home')->with('warning', 'Anda Tidak Punya Ijin Untuk Mengedit Data Ini');
+        }
         $categories = BookCategory::whereNull('deleted_at')->get();
 
         return view('admin.pages.edit-book', compact('book', 'categories'));
@@ -121,13 +123,55 @@ class BookController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\UpdateBookRequest  $request
-     * @param  \App\Models\Book  $book
+     * @param  \App\Http\Requests\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateBookRequest $request, Book $book)
+    public function update(Request $request, $book)
     {
-        //
+
+        $book = Book::find($book);
+        if (!Gate::allows('update-book', $book)) {
+            return to_route('home')->with('warning', 'Anda Tidak Punya Ijin Untuk Mengupdate Data Ini');
+        }
+
+        $validated = $request->validate([
+            'title' => 'required|string|max:255|unique:books,title,' . $book->id,
+            'quantity' => 'required|integer',
+            'category_id' => 'required',
+            'cover' => 'mimes:jpg,jpeg,png,bmp,tiff|max:4096',
+            'file' => 'mimes:pdf|max:10000',
+            'description' => 'required'
+        ]);
+
+        if ($request->file('cover')) {
+
+            if (file_exists(public_path('uploads/book-cover/' . $book->cover))) {
+                unlink(public_path('uploads/book-cover/' . $book->cover));
+            }
+            $cover = time() . '.' . $request->cover->extension();
+
+            $request->cover->move(public_path('uploads/book-cover'), $cover);
+
+            $validated['cover'] = $cover;
+        }
+
+        if ($request->file('file')) {
+            if (file_exists(public_path('uploads/book-file/' . $book->file))) {
+                unlink(public_path('uploads/book-file/' . $book->file));
+                /*
+                    Delete Multiple files this way
+                    Storage::delete(['upload/test.png', 'upload/test2.png']);
+                */
+            }
+            $file = time() . '.' . $request->file->extension();
+
+            $request->file->move(public_path('uploads/book-file'), $file);
+            $validated['cover'] = $file;
+        }
+
+        $book->slug = '';
+        $book->update($validated);
+        return to_route('book.index')->with('success', trans('response.success.update', ['data' => 'Data Buku']));
     }
 
     /**
@@ -139,6 +183,10 @@ class BookController extends Controller
     public function destroy($book)
     {
         $book = Book::find($book);
+
+        if (!Gate::allows('delete-book', $book)) {
+            return to_route('home')->with('warning', 'Anda Tidak Punya Ijin Untuk Menghapus Data Ini');
+        }
 
         if (file_exists(public_path('uploads/book-cover/' . $book->cover))) {
             unlink(public_path('uploads/book-cover/' . $book->cover));
